refactor: 메서드 책임 분리
